move locales en and en-US into the core
