+++ lib/index-basic.js
@@ -15,10 +15,4 @@ Plotly.register([
     require('./pie')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
