+++ lib/index-cartesian.js
@@ -23,10 +23,4 @@ Plotly.register([
     require('./violin')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
