+++ lib/index-finance.js
@@ -18,10 +18,4 @@ Plotly.register([
     require('./candlestick')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
