+++ lib/index-gl2d.js
@@ -18,10 +18,4 @@ Plotly.register([
     require('./parcoords')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
