+++ lib/index-gl3d.js
@@ -16,10 +16,4 @@ Plotly.register([
     require('./mesh3d')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
