+++ lib/index-mapbox.js
@@ -14,10 +14,4 @@ Plotly.register([
     require('./scattermapbox')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en'),
-    require('./locale-en-us')
-]);
-
 module.exports = Plotly;
