+++ lib/index.js
@@ -71,10 +71,4 @@ Plotly.register([
     require('./calendars')
 ]);
 
-// locales
-Plotly.register([
-    require('./locale-en.js'),
-    require('./locale-en-us.js')
-]);
-
 module.exports = Plotly;
